% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orReferenceTree}
\alias{orReferenceTree}
\title{O'Reilly et al. reference tree}
\format{A single phylogenetic tree saved as an object of class \code{phylo}}
\source{
\insertRef{OReilly2016}{Quartet}
}
\usage{
orReferenceTree
}
\description{
The tree topology used to generate the matrices in O'Reilly \emph{et al.} 2016
}
\examples{
  library('ape') # Contains tree plotting functions
  data(orReferenceTree)
  plot(orReferenceTree)

}
\references{
\insertRef{OReilly2016}{Quartet}
}
\keyword{datasets}
